TFP-6341 oppdager og logger tilfelle av falsk identitet
