+++ domenetjenester/person/src/main/java/no/nav/foreldrepenger/domene/person/pdl/PersonBasisTjeneste.java
@@ -8,10 +8,6 @@ import java.util.Optional;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 
-import no.nav.pdl.FolkeregisteridentifikatorResponseProjection;
-
-import no.nav.pdl.FolkeregisterpersonstatusResponseProjection;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -30,10 +26,9 @@ import no.nav.pdl.AdressebeskyttelseGradering;
 import no.nav.pdl.AdressebeskyttelseResponseProjection;
 import no.nav.pdl.Doedsfall;
 import no.nav.pdl.DoedsfallResponseProjection;
-import no.nav.pdl.FalskIdentitetIdentifiserendeInformasjonResponseProjection;
-import no.nav.pdl.FalskIdentitetResponseProjection;
 import no.nav.pdl.Foedselsdato;
 import no.nav.pdl.FoedselsdatoResponseProjection;
+import no.nav.pdl.FolkeregisteridentifikatorResponseProjection;
 import no.nav.pdl.HentPersonQueryRequest;
 import no.nav.pdl.Kjoenn;
 import no.nav.pdl.KjoennResponseProjection;
@@ -41,7 +36,6 @@ import no.nav.pdl.KjoennType;
 import no.nav.pdl.NavnResponseProjection;
 import no.nav.pdl.Person;
 import no.nav.pdl.PersonResponseProjection;
-import no.nav.pdl.PersonnavnResponseProjection;
 
 @ApplicationScoped
 public class PersonBasisTjeneste {
@@ -64,13 +58,13 @@ public class PersonBasisTjeneste {
         var query = new HentPersonQueryRequest();
         query.setIdent(aktørId.getId());
         var projection = new PersonResponseProjection()
-            .falskIdentitet(new FalskIdentitetResponseProjection().erFalsk())
+            .folkeregisteridentifikator(new FolkeregisteridentifikatorResponseProjection().identifikasjonsnummer())
             .navn(new NavnResponseProjection().fornavn().mellomnavn().etternavn())
             .adressebeskyttelse(new AdressebeskyttelseResponseProjection().gradering());
 
         var person = pdlKlient.hentPerson(ytelseType, query, projection);
 
-        sjekkFalskIdentitet(person, ytelseType, aktørId);
+        pdlKlient.sjekkPersonFalskIdentitet(ytelseType, aktørId, person.getFolkeregisteridentifikator());
 
         return new PersoninfoVisning(aktørId, personIdent, mapNavnVisning(person, aktørId), getDiskresjonskode(person));
     }
@@ -79,7 +73,7 @@ public class PersonBasisTjeneste {
         var query = new HentPersonQueryRequest();
         query.setIdent(aktørId.getId());
         var projection = new PersonResponseProjection()
-            .falskIdentitet(new FalskIdentitetResponseProjection().erFalsk())
+            .folkeregisteridentifikator(new FolkeregisteridentifikatorResponseProjection().identifikasjonsnummer())
             .navn(new NavnResponseProjection().fornavn().mellomnavn().etternavn())
             .foedselsdato(new FoedselsdatoResponseProjection().foedselsdato())
             .doedsfall(new DoedsfallResponseProjection().doedsdato())
@@ -88,7 +82,7 @@ public class PersonBasisTjeneste {
 
         var person = pdlKlient.hentPerson(ytelseType, query, projection);
 
-        sjekkFalskIdentitet(person, ytelseType, aktørId);
+        pdlKlient.sjekkPersonFalskIdentitet(ytelseType, aktørId, person.getFolkeregisteridentifikator());
 
         var fødselsdato = person.getFoedselsdato().stream()
             .map(Foedselsdato::getFoedselsdato)
@@ -159,55 +153,6 @@ public class PersonBasisTjeneste {
             .medNavBrukerKjønn(mapKjønn(person))
             .build();
         return person.getKjoenn().isEmpty() ? Optional.empty() : Optional.of(kjønn);
-    }
-
-    private void sjekkFalskIdentitet(Person person, FagsakYtelseType ytelseType, AktørId aktørId) {
-        if (person.getFalskIdentitet() != null && person.getFalskIdentitet().getErFalsk()) {
-            var query = new HentPersonQueryRequest();
-            query.setIdent(aktørId.getId());
-            var projection = new PersonResponseProjection()
-                .folkeregisteridentifikator(new FolkeregisteridentifikatorResponseProjection().status().identifikasjonsnummer())
-                .folkeregisterpersonstatus(new FolkeregisterpersonstatusResponseProjection().status())
-                .falskIdentitet(new FalskIdentitetResponseProjection().rettIdentitetErUkjent().rettIdentitetVedIdentifikasjonsnummer()
-                    .rettIdentitetVedOpplysninger(new FalskIdentitetIdentifiserendeInformasjonResponseProjection().kjoenn().foedselsdato()
-                        .personnavn(new PersonnavnResponseProjection().fornavn().mellomnavn().etternavn()).statsborgerskap()));
-            var falskIdentitetPerson = pdlKlient.hentPerson(ytelseType, query, projection);
-
-            if (falskIdentitetPerson.getFolkeregisteridentifikator() != null) {
-                var identifikatorer = falskIdentitetPerson.getFolkeregisteridentifikator().stream()
-                    .map(p -> p.getIdentifikasjonsnummer().substring(p.getIdentifikasjonsnummer().length() - 5) + " " + p.getStatus())
-                    .toList();
-                LOG.info("Falsk identitet aktør {} har identer {}", aktørId, identifikatorer);
-            }
-            if (falskIdentitetPerson.getFolkeregisterpersonstatus() != null) {
-                var statuser = falskIdentitetPerson.getFolkeregisterpersonstatus().stream()
-                    .map(p -> p.getStatus())
-                    .toList();
-                LOG.info("Falsk identitet aktør {} har statuser {}", aktørId, statuser);
-            }
-            if (Objects.equals(falskIdentitetPerson.getFalskIdentitet().getRettIdentitetErUkjent(), Boolean.TRUE)) {
-                LOG.info("Falsk identitet aktør {} har rettIdentitetErUkjent", aktørId);
-            } else if (falskIdentitetPerson.getFalskIdentitet().getRettIdentitetVedIdentifikasjonsnummer() != null) {
-                LOG.info("Falsk identitet aktør {} har rettIdentitetVedIdentifikasjonsnummer {}", aktørId, falskIdentitetPerson.getFalskIdentitet().getRettIdentitetVedIdentifikasjonsnummer());
-            } else if (falskIdentitetPerson.getFalskIdentitet().getRettIdentitetVedOpplysninger() != null) {
-                // .personnavn(new PersonnavnResponseProjection().fornavn().mellomnavn().etternavn()).statsborgerskap())
-                var kjønn = falskIdentitetPerson.getFalskIdentitet().getRettIdentitetVedOpplysninger().getKjoenn();
-                var statsborgerskap = falskIdentitetPerson.getFalskIdentitet().getRettIdentitetVedOpplysninger().getStatsborgerskap();
-                var fødselsdato = falskIdentitetPerson.getFalskIdentitet().getRettIdentitetVedOpplysninger().getFoedselsdato();
-                var navn = Optional.ofNullable(falskIdentitetPerson.getFalskIdentitet().getRettIdentitetVedOpplysninger().getPersonnavn())
-                    .map(p -> Optional.ofNullable(p.getFornavn()).orElse("UtenFN")
-                        + leftPad(Optional.ofNullable(p.getMellomnavn()).orElse("UtenMN"))
-                        + leftPad(Optional.ofNullable(p.getEtternavn()).map(e -> "ETTERN").orElse("UtenEN")))
-                    .orElse("UtenNavn");
-                LOG.info("Falsk identitet aktør {} har rettIdentitetVedOpplysninger navn {} fdato {} kjønn {} statsborger {}", aktørId, navn, fødselsdato, kjønn, statsborgerskap);
-            } else {
-                LOG.info("Falsk identitet aktør {} mangler info om rett identitet", aktørId);
-            }
-        }
-    }
-
-    private static String leftPad(String navn) {
-        return Optional.ofNullable(navn).map(n -> " " + navn).orElse("");
     }
 
     private Diskresjonskode getDiskresjonskode(Person person) {
